Move progress from music to labs folder
